MiddlewareForFANUC-zRoKi/v1.1_zFANUC_Middleware/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==1.15.0
dnspython3==1.15.0
Flask==1.1.2
Flask-MQTT==1.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
paho-mqtt==1.6.1
Pillow==8.4.0
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.3
Werkzeug==1.0.1
